+++ src/landscape/open_street_map/path/mod.rs
@@ -1,6 +1,8 @@
+#[cfg(test)]
+mod tests;
+
 use crate::{landscape::CoordinatePoint, train::Direction};
 use bevy::prelude::*;
-use rand::{thread_rng, Rng};
 use serde::{Deserialize, Serialize};
 
 pub type PathId = (i64, i64);
@@ -12,7 +14,7 @@ pub struct Path {
     pub start_coords: CoordinatePoint,
     pub end_coords: CoordinatePoint,
     pub forward_connections: Vec<(PathId, Direction)>,
-    pub backwards_connections: Vec<(PathId, Direction)>,
+    pub backward_connections: Vec<(PathId, Direction)>,
 }
 
 impl Path {
@@ -21,26 +23,13 @@ impl Path {
     }
 }
 
-// TODO
-#[allow(dead_code)]
+/*
 impl Path {
     fn possible_connections_by_direction(&self, direction: Direction) -> &Vec<(PathId, Direction)> {
         match direction {
             Direction::Forward => &self.forward_connections,
-            Direction::Backward => &self.backwards_connections,
+            Direction::Backward => &self.backward_connections,
         }
-    }
-
-    fn random_connection_by_direction(&self, direction: Direction) -> Option<&(PathId, Direction)> {
-        let possible = self.possible_connections_by_direction(direction);
-
-        if possible.is_empty() {
-            return None;
-        }
-
-        let mut random = thread_rng();
-        let index = random.gen_range(0..possible.len());
-
-        Some(&possible[index])
     }
 }
+*/
